[refactor] move /AddCase folder our of /QRCodeScanner folder.
