+++ src/app/(BottomTabNavigation)/AllCases/AddCase/styles.ts
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import { colors } from '../../../../../styles/colors';
+import { colors } from '../../../../styles/colors';
 
 export default StyleSheet.create({
   container: {
